pyproject.toml
[tool.poetry]
name = "catflap"
version = "0.1.0"
description = "YouTube Live Chat stream proxy"
authors = ["yay4ya <yay4ya.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytchat = "^0.0.8"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.5.3"
mypy = "^0.782"

[tool.poetry.scripts]
catflap = "catflap.__main__:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
